File 'game.ts' moved to 'game/game.ts'
